third_party/blink/web_tests/platform/linux-chrome/external/wpt/fetch/stale-while-revalidate/fetch.any.sharedworker-expected.txt

This is a testharness.js-based test.
[FAIL] Second fetch returns same response
  assert_equals: expected "ielxkyhfptrxwhzjbtap" but got "hipqbmwrlskkazdrrbsv"
Harness: the test ran to completion.
